Bug:49565 - create manifest data for lambda packages (#67)
* Bug:49565 - create manifest data for lambda packages

* Changed last modified time to the time building the package

* Minor changes

+++ lambda/package/managedeni_pkg.sh
@@ -1,10 +1,15 @@
 #!/bin/sh
-rm ../managed_eni.zip
-cp custom_resource_base.zip managed_eni.zip
 
-mv managed_eni.zip ../
+[ -e ../managed_eni.zip ] && rm ../managed_eni.zip
+[ -e ../version.txt ] && rm ../version.txt
+
+echo `date` > ../version.txt
+
+cp custom_resource_base.zip ../managed_eni.zip
+
 cd ..
-zip -g managed_eni.zip managedeni.py
-zip -g managed_eni.zip utils.py
+zip -g managed_eni.zip managedeni.py utils.py version.txt
+
+rm version.txt
 
 chmod 777 managed_eni.zip
